revert

+++ backend_py/libs/services/src/webviz_services/pdm_access/_calculate_totals_from_daily.py
@@ -59,7 +59,7 @@ def calculate_total_injection_from_daily(
     api_results: list[dict], start_date: str, end_date: str
 ) -> list[WellInjectionData]:
     """Calculate total injection per well from daily injection data."""
-    polars_df = pl.DataFrame(api_results, schema_overrides={_INJCOLUMNS.WB_INJ_VOL: pl.Float64})
+    polars_df = pl.DataFrame(api_results, schema_overrides={_INJCOLUMNS.WB_INJ_VOL: pl.Float32})
 
     # Group per well and sum injection volumes by type
     grouped_df = (
